requirements.txt
python-telegram-bot==20.0
TA-Lib==0.4.28
ccxt==3.2.43
numpy==1.21.6
python-dotenv

africastalking==1.2.5
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aioredis==1.3.1
aiosignal==1.3.2
amqp==5.1.1
anyio==4.9.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.1.0
binance-connector==3.12.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
channels==2.3.1
channels-redis==2.4.2
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
contextlib2==21.6.0
crispy-bootstrap4==2024.1
cron-descriptor==1.4.0
cryptography==43.0.1
cssselect2==0.7.0
daphne==2.5.0
dateparser==1.2.1
distlib==0.3.7
Django==4.2
django-appconf==1.0.5
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-channels==0.7.0
django-ckeditor==6.6.1
django-cors-headers==4.0.0
django-crispy-forms==2.2
django-extensions==3.2.3
django-imagekit==4.1.0
django-js-asset==2.1.0
django-timezone-field==5.1
env==0.1.0
filelock==3.12.4
frozenlist==1.5.0
h11==0.14.0
hiredis==3.0.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.4
incremental==24.7.2
kombu==5.3.1
lxml==4.9.3
msgpack==0.6.2
multidict==6.2.0
mysqlclient==2.2.4
numpy==1.23.5
oauthlib==3.2.2
oscrypto==1.3.0
pandas==2.2.3
pilkit==2.0
Pillow==10.0.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
propcache==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
pycryptodome==3.22.0
pyHanko==0.19.0
pyhanko-certvalidator==0.23.0
pyOpenSSL==24.2.1
pypdf==3.12.0
pypng==0.20220715.0
python-bidi==0.4.2
python-binance==1.0.28
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.1.0
python-telegram-bot==22.0
pytz==2025.2
PyYAML==6.0
qrcode==7.4.2
redis==4.6.0
regex==2024.11.6
reportlab==3.6.13
requests==2.31.0
requests-oauthlib==1.3.1
schema==0.7.5
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
svglib==1.5.1
ta==0.11.0
TA-Lib @ file:///C:/Users/Administrator/Downloads/TA_Lib-0.4.28-cp311-cp311-win_amd64.whl#sha256=5bc83ddd114b65da7f8c2ec386c5052922b2a3947ba22ab91dc9e8cc2c5467dc
tinycss2==1.2.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.1
urllib3==2.0.3
vboxapi==1.0
vine==5.0.0
virtualenv==20.24.5
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
whitenoise==6.5.0
xhtml2pdf==0.2.11
yarl==1.18.3
zope.interface==7.1.0
